Add XmlAppendingTransformerTest
