+++ src/functionalTest/kotlin/com/github/jengelman/gradle/plugins/shadow/transformers/TransformersTest.kt
@@ -48,48 +48,6 @@ class TransformersTest : BaseTransformerTest() {
     commonAssertions()
   }
 
-  @Test
-  fun appendXmlFiles() {
-    val propertiesXml = "properties.xml"
-    val xmlContent = """
-      <?xml version="1.0" encoding="UTF-8"?>
-      <!DOCTYPE properties SYSTEM "http://java.sun.com/dtd/properties.dtd">
-      <properties version="1.0">
-        <entry key="%s">%s</entry>
-      </properties>
-    """.trimIndent()
-
-    val xml1 = buildJar("xml1.jar") {
-      insert(propertiesXml, xmlContent.format("key1", "val1"))
-    }
-    val xml2 = buildJar("xml2.jar") {
-      insert(propertiesXml, xmlContent.format("key2", "val2"))
-    }
-
-    projectScriptPath.appendText(
-      transform<XmlAppendingTransformer>(
-        shadowJarBlock = fromJar(xml1, xml2),
-        transformerBlock = """
-          resource = 'properties.xml'
-        """.trimIndent(),
-      ),
-    )
-
-    run(shadowJarTask)
-
-    val content = outputShadowJar.use { it.getContent(propertiesXml) }.trimIndent()
-    assertThat(content).isEqualTo(
-      """
-        <?xml version="1.0" encoding="UTF-8"?>
-        <!DOCTYPE properties SYSTEM "http://java.sun.com/dtd/properties.dtd">
-        <properties version="1.0">
-          <entry key="key1">val1</entry>
-          <entry key="key2">val2</entry>
-        </properties>
-      """.trimIndent(),
-    )
-  }
-
   @Issue("https://github.com/GradleUp/shadow/issues/82")
   @Test
   fun shadowManifestLeaksToJarManifest() {
